requirements.txt
flask==3.0.3
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
textblob==0.18.0
requests==2.32.3
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1